refactor(body): add information personal and correction styles
