+++ src/layouts/developer/components/Header/index.js
@@ -23,7 +23,7 @@ import Grid from "@mui/material/Grid";
 import Tab from "@mui/material/Tab";
 import Tabs from "@mui/material/Tabs";
 // Images
-import burceMars from "assets/images/profile-logo-ntback.png";
+import burceMars from "assets/images/logo-v2.png";
 // Vision UI Dashboard React base styles
 import breakpoints from "assets/theme/base/breakpoints";
 import VuiAvatar from "components/VuiAvatar";
@@ -69,8 +69,9 @@ function Header() {
       <DashboardNavbar light />
       <Card
         sx={{
-          px: 3,
+          px: 2,
           mt: 2,
+          py: 1,
         }}
       >
         <Grid
@@ -108,7 +109,7 @@ function Header() {
               src={burceMars}
               alt="profile-image"
               variant="rounded"
-              size="xl"
+              size="lg"
               shadow="sm"
             />
           </Grid>
